added a genomescan check in the movesamplesheets

+++ bin/moveAndCheckSamplesheets.sh
@@ -299,11 +299,18 @@ do
 	#
 	if [[ "${projectSamplesheet}" == "true" ]]
 	then
-		firstStepOfPipeline="NGS_DNA"
-		perl -p -e "s|${REPLACEDPIPELINECOLUMN}|${firstStepOfPipeline}|" "${samplesheet}" > "${samplesheet}.tmp"
-		mv "${samplesheet}.tmp" "${samplesheet}"
-		log4Bash 'INFO' "${LINENO}" "${FUNCNAME:-main}" '0' "The samplesheet is a project samplesheet (no NGS_Demultiplexing); firstStepOfPipeline was set to ${firstStepOfPipeline}."
+		if [[ "${REPLACEDPIPELINECOLUMN}" == *"GENOMESCAN"* ]]
+		then
+			firstStepOfPipeline=''
+			log4Bash 'INFO' "${LINENO}" "${FUNCNAME:-main}" '0' "The samplesheet is a GENOMESCAN project samplesheet, the first step of the pipeline will be set to an empty string (samplesheet will be put in correct bucket in a later stage of the pipeline)."
+		else
+			firstStepOfPipeline="NGS_DNA"
+			perl -p -e "s|${REPLACEDPIPELINECOLUMN}|${firstStepOfPipeline}|" "${samplesheet}" > "${samplesheet}.tmp"
+			mv "${samplesheet}.tmp" "${samplesheet}"
+			log4Bash 'INFO' "${LINENO}" "${FUNCNAME:-main}" '0' "The samplesheet is a project samplesheet (no NGS_Demultiplexing); firstStepOfPipeline was set to ${firstStepOfPipeline}."
+		fi
 	fi
+	log4Bash 'TRACE' "${LINENO}" "${FUNCNAME:-main}" '0' "first step of the pipeline:[${firstStepOfPipeline}]."
 	# shellcheck disable=SC2153
 	samplesheetDestination="${HOSTNAME_TMP}:/groups/${GROUP}/${SCR_LFS}/Samplesheets/${firstStepOfPipeline}/"
